# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.envelope:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  handler.errorEnvelope:
    properties:
      errors:
        items:
          $ref: '#/definitions/handler.fieldError'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  handler.fieldError:
    properties:
      detail:
        type: string
      field:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://github.com/404nfidv2/go-nuxt-starter-kit
  description: REST API for the Go-Nuxt Starter Kit (auth, profile, notifications,
    admin).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go-Nuxt Starter Kit API
  version: "1.0"
paths:
...
  /auth/change-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Passwords
        in: body
        name: body
        required: true
        schema:
          properties:
            current_password:
              type: string
            new_password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Change password (requires auth)
      tags:
      - auth
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registered email
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Request a password reset link
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.envelope'
            - properties:
                data:
                  properties:
                    access_token:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Login with email and password
      tags:
      - auth
...
  /auth/refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.envelope'
            - properties:
                data:
                  properties:
                    access_token:
                      type: string
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Refresh access token using HTTP-only refresh cookie
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
            first_name:
              type: string
            last_name:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Register a new user
      tags:
      - auth
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email address
        in: body
        name: body
        required: true
        schema:
          properties:
            email:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Resend email verification link
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset payload
        in: body
        name: body
        required: true
        schema:
          properties:
            password:
              type: string
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Reset password using token from email
      tags:
      - auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: Verification token
        in: body
        name: body
        required: true
        schema:
          properties:
            token:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      summary: Verify email address
      tags:
      - auth
...
  /notifications:
    get:
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: List notifications for the current user
      tags:
      - notifications
  /notifications/{id}/read:
    patch:
      parameters:
      - description: Notification UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Mark a notification as read
      tags:
      - notifications
  /notifications/read-all:
    patch:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Mark all notifications as read
      tags:
      - notifications
  /notifications/unread-count:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.envelope'
            - properties:
                data:
                  properties:
                    unread_count:
                      type: integer
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Get unread notification count
      tags:
      - notifications
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"